Colony is an application perse and there is no reason to create a component space inside this application. Moved the application component to the it's root.
