+++ socialspider/colony/conf/iflux.ini
@@ -15,7 +15,7 @@ data.connection.instance.session.type=redis
 
 [Component]
 
-components.s2_colony=socialspider.colony.components.colony.ColonyComponent
+components.s2_colony=socialspider.colony.ColonyComponent
 
 components.s2_colony.enabled=true
 #components.admin.enabled=true
